organize raster mod
